Fixed a MAJOR problem with randomization. Ok it could still probably be better, but at least it's now not totally worthless as it was (this is one of several reasons you should think twice about using time() as a seed even in non cryptographic sitations).
Also, while this could still probably be better, at least this time I'm only calling random() a couple of times in total (till the kilocores come out ohfuk..

And at least it's all in the same thread for this usage. Not sure if I should stick with it, or use something else like /dev/random. I don't like that it's not very portable for one.